Update README.md; Update Install Script

+++ test/util-test.h
@@ -7,7 +7,7 @@ std::string bytes2hex(const uint8_t * input, size_t input_len){
     static const char *sha2_hex_digits = "0123456789abcdef";
     std::unique_ptr<char[]> output_hex(new char [input_len * 2 + 1]);
     const uint8_t *d = input;
-    for (int i = 0; i < input_len; i++) {
+    for (size_t i = 0; i < input_len; i++) {
         output_hex[i * 2] = sha2_hex_digits[(*d & 0xf0) >> 4];
         output_hex[i * 2 + 1] = sha2_hex_digits[*d & 0x0f];
         d++;
